refactor: useReducer 훅 분리
- useReducer 훅 분리
