+++ src/components/pages/mypage/MyPageInfoTap.tsx
@@ -1,70 +1,29 @@
 "use client"
 
-import { useReducer } from "react"
-
-import { IAction, IInitialState } from "@/types/mypage/mypage"
+import useTapHook from "@/hooks/mypage/useTapHook"
 
 import MyPageInfoTapButton from "./MyPageInfoTapButton"
 import MyPageInfoWrapper from "./MyPageInfoWrapper"
 
-const initialState = {
-  myMeeting: true,
-  myReview: false,
-  myOwnMeeting: false,
-}
-
-const reducer = (state: IInitialState, action: IAction) => {
-  switch (action.type) {
-    case "myMeeting":
-      return {
-        ...state,
-        myMeeting: true,
-        myReview: false,
-        myOwnMeeting: false,
-      }
-    case "myReview":
-      if (action.isReviewed) {
-        return {
-          ...state,
-          myMeeting: false,
-          myReview: true,
-          myOwnMeeting: false,
-          isReviewed: action.isReviewed,
-        }
-      }
-      return {
-        ...state,
-        myMeeting: false,
-        myReview: true,
-        myOwnMeeting: false,
-      }
-    case "myOwnMeeting":
-      return {
-        ...state,
-        myMeeting: false,
-        myReview: false,
-        myOwnMeeting: true,
-      }
-    default:
-      return state
-  }
-}
-
 const animatedBottomClassName = (state: string) => {
+  const myMeetingClass = "w-[66px] -translate-x-[1px] md:w-[74px]"
+  const myReviewClass = "w-[54px] translate-x-[76px] md:w-[62px] md:translate-x-[85px]"
+  const myOwnMeetingClass = "w-[82px] translate-x-[140px] md:w-[94px] md:translate-x-[155px]"
+
   switch (state) {
     case "myMeeting":
-      return "w-[66px] -translate-x-[1px] md:w-[74px]"
+      return myMeetingClass
     case "myReview":
-      return "w-[54px] translate-x-[76px] md:w-[62px] md:translate-x-[85px]"
+      return myReviewClass
     case "myOwnMeeting":
-      return "w-[82px] translate-x-[140px] md:w-[94px] md:translate-x-[155px]"
+      return myOwnMeetingClass
     default:
-      return "w-[66px] -translate-x-[1px] md:w-[74px]"
+      return myMeetingClass
   }
 }
 
 const MyPageInfoTap = () => {
-  const [tapState, dispatch] = useReducer(reducer, initialState)
+  const { tapState, dispatch } = useTapHook()
   const [[dataFetchingKey]] = Object.entries(tapState).filter((state) => {
     return state[1] === true && state[0] !== "isReviewed"
   })
